refactor(frontend): remove content loaders, cleanup
