+++ apps/web/src/serlo-editor-integration/serlo-editor.tsx
@@ -5,6 +5,7 @@ import { editStrings as editStringsEn } from '@editor/i18n/strings/en/edit'
 import { editorLearnerEvent } from '@editor/plugin/helpers/editor-learner-event'
 import { editorPlugins } from '@editor/plugin/helpers/editor-plugins'
 import { editorRenderers } from '@editor/plugin/helpers/editor-renderer'
+import { AnyEditorDocument } from '@editor/types/editor-plugins'
 import { TemplatePluginType } from '@editor/types/template-plugin-type'
 import { SerloOnlyFeaturesContext } from '@editor/utils/serlo-extra-context'
 import dynamic from 'next/dynamic'
@@ -12,7 +13,7 @@ import { mergeDeepRight } from 'ramda'
 import { type ReactNode } from 'react'
 
 import { ArticleAddModal } from './components/article-add-modal/article-add-modal'
-import { ContentLoaders } from './components/content-loaders/content-loaders'
+import { ExternalRevisionLoader } from './components/external-revision-loader'
 import { SaveButton } from './components/save-button'
 import { createPlugins } from './create-plugins'
 import { createRenderers } from './create-renderers'
@@ -52,6 +53,8 @@ export function SerloEditor({
   const editString =
     lang === 'de' ? mergeDeepRight(editStringsEn, editStringsDe) : editStringsEn
 
+  const isNewEntity = !(initialState.state as AnyEditorDocument)?.id
+
   return (
     <EditStringsProvider value={editString}>
       <SerloOnlyFeaturesContext.Provider
@@ -59,27 +62,15 @@ export function SerloEditor({
       >
         <Editor initialState={initialState}>
           <SaveButton onSave={onSave} isInTestArea={isInTestArea} />
-          {renderContentLoaders()}
+          {isNewEntity ? (
+            <ExternalRevisionLoader
+              templateType={initialState.plugin as TemplatePluginType}
+            />
+          ) : null}
+
           {children}
         </Editor>
       </SerloOnlyFeaturesContext.Provider>
     </EditStringsProvider>
   )
-
-  function renderContentLoaders() {
-    const templateType = initialState.plugin as TemplatePluginType
-    if (!pluginsWithContentLoaders.includes(templateType)) return null
-
-    return <ContentLoaders templateType={templateType} />
-  }
 }
-
-const pluginsWithContentLoaders = [
-  TemplatePluginType.Applet,
-  TemplatePluginType.Article,
-  TemplatePluginType.Course,
-  TemplatePluginType.Event,
-  TemplatePluginType.TextExercise,
-  TemplatePluginType.TextExerciseGroup,
-  TemplatePluginType.Video,
-]
